Centralize business logic within the TodoProvider
