+++ react-native/SampleProfileApp/src/todo-app-use-context/context/TodoProvider.js
@@ -8,8 +8,32 @@ export const TodoProvider = ({ children }) => {
         new Todo(2, "Item # 2"),
     ]);
 
+    const addTodo = (text) => {
+        setTodos([...todos, new Todo(todos.length + 1, text)]);
+    };
+
+    const markComplete = (todo) => {
+        todo.markDone();
+        setTodos([...todos]);
+    };
+
+    const deleteTodo = (todo) => {
+        setTodos(todos.filter(item => item.id !== todo.id));
+    };
+
+    const editTodo = (id, newText) => {
+        setTodos(
+            todos.map(item => {
+                if (item.id === id) {
+                    item.updateText(newText);
+                }
+                return item;
+            })
+        );
+    };
+
     return (
-        <TodoContext.Provider value={{ todos, setTodos }}>
+        <TodoContext.Provider value={{ todos, addTodo, markComplete, deleteTodo, editTodo }}>
             {children}
         </TodoContext.Provider>
     );
